feat: member, auth 연동

+++ backend/src/main/java/consolelog/post/service/PostService.java
@@ -1,5 +1,11 @@
 package consolelog.post.service;
 
+import consolelog.auth.dto.AuthInfo;
+import consolelog.auth.exception.AuthorizationException;
+import consolelog.auth.service.AuthService;
+import consolelog.member.domain.Member;
+import consolelog.member.exception.MemberNotFoundException;
+import consolelog.member.repository.MemberRepository;
 import consolelog.post.domain.Post;
 import consolelog.post.dto.request.NewPostRequest;
 import consolelog.post.dto.request.PostUpdateRequest;
@@ -17,13 +23,18 @@ import java.util.Optional;
 @Transactional(readOnly = true)
 public class PostService {
     private final PostRepository postRepository; // final로 선언하면 생성자에서만 값을 할당할 수 있음
+    private final MemberRepository memberRepository;
+    private final AuthService authService;
 
-    public PostService(PostRepository postRepository) { // 생성자 주입
+    public PostService(PostRepository postRepository, MemberRepository memberRepository, AuthService authService) { // 생성자 주입
         this.postRepository = postRepository;  // 생성자를 통해 PostRepository를 주입받음
+        this.memberRepository = memberRepository;
+        this.authService = authService;
     }
 
 
     public PostResponse findPost(Long postId) {// post_id 게시글 조회
+
         Post findPost = findPostById(postId);
         return PostResponse.of(findPost);
     }
@@ -31,6 +42,11 @@ public class PostService {
     private Post findPostById(Long postId) {
         return postRepository.findById(postId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id=" + postId));
+    }
+
+    public PagePostResponse findPostsByPage(Long lastPostId, Pageable pageable) {
+        Slice<Post> posts = postRepository.findByIdIsLessThanOrderByIdDesc(lastPostId, pageable);
+        return PagePostResponse.of(posts);
     }
 
     public PostResponse findTitle(String title) {// post_title 게시글 조회
@@ -44,39 +60,50 @@ public class PostService {
     }
 
     @Transactional
-    public Long addPost(NewPostRequest newPostRequest) {
-        Post post = createPost(newPostRequest);
+    public Long addPost(NewPostRequest newPostRequest, AuthInfo authInfo) {
+        Member member = findMember(authInfo);
+        Post post = createPost(newPostRequest, member);
         Optional<Post> savedPost = Optional.of(postRepository.save(post));
         return savedPost.orElseThrow(() -> new IllegalArgumentException("x")).getId();
     }
 
-    private Post createPost(NewPostRequest newPostRequest) {
+    private Member findMember(AuthInfo authInfo) {
+        return memberRepository.findById(authInfo.getId())
+                .orElseThrow(MemberNotFoundException::new);
+    }
+
+    private Post createPost(NewPostRequest newPostRequest, Member member) {
         return Post.builder()
                 .title(newPostRequest.getTitle())
                 .content(newPostRequest.getContent())
+                .member(member)
                 .build();
     }
 
     @Transactional
-    public void updatePost(Long id, PostUpdateRequest postUpdateRequest) {
+    public void updatePost(Long id, PostUpdateRequest postUpdateRequest, AuthInfo authInfo) {
         Post post = findPostById(id);
+        validateOwner(authInfo, post);
         post.updateTitle(postUpdateRequest.getTitle());
         post.updateContent(postUpdateRequest.getContent());
     }
 
-    @Transactional
-    public void deletePost(Long id) {
-        Post post = findPostById(id);
-        postRepository.delete(post);
 
+    @Transactional
+    public void deletePost(Long id, AuthInfo authInfo) {
+        Post post = findPostById(id);
+        validateOwner(authInfo, post);
+        postRepository.delete(post);
         //postRepository.deleteById(id);
     }
 
 
     //모든 게시물 조회 - 페이징 처리
-    public PagePostResponse findPostsByPage(Long lastPostId, Pageable pageable) {
-        Slice<Post> posts = postRepository.findByIdIsLessThanOrderByIdDesc(lastPostId, pageable);
-        return PagePostResponse.of(posts);
+
+    private void validateOwner(AuthInfo authInfo, Post post) {
+        if (!post.isOwner(authInfo.getId())) {
+            throw new AuthorizationException();
+        }
     }
 
 }
